test(snapshot): dedupe table-root delta path helper
